Removed vectorpipe from the code
