+++ geopyspark-backend/geotrellis/src/main/scala/geopyspark/geotrellis/protobufs/Implicits.scala
@@ -12,3 +12,4 @@ trait Implicits extends CRSProtoBuf
     with TileProtoBuf
     with MultibandTileProtoBuf
     with TupleProtoBuf
+    with FeatureProtoBuf
