--- geopyspark-backend/vectorpipe/build.sbt
@@ -1,25 +1,0 @@
-name := "vectorpipe"
-
-libraryDependencies ++= Seq(
-  "org.apache.spark"            %% "spark-hive"            % "2.3.0" % "provided",
-  "org.locationtech.geotrellis" %% "geotrellis-s3"         % Version.geotrellis,
-  "org.locationtech.geotrellis" %% "geotrellis-s3-testkit" % Version.geotrellis,
-  "org.locationtech.geotrellis" %% "geotrellis-spark"      % Version.geotrellis,
-  "org.typelevel"               %% "cats"                  % "0.9.0",
-  "com.azavea" %% "vectorpipe" % "0.3.0"
-)
-
-assemblyMergeStrategy in assembly := {
-  case "reference.conf" => MergeStrategy.concat
-  case "application.conf" => MergeStrategy.concat
-  case "META-INF/MANIFEST.MF" => MergeStrategy.discard
-  case "META-INF\\MANIFEST.MF" => MergeStrategy.discard
-  case "META-INF/ECLIPSEF.RSA" => MergeStrategy.discard
-  case "META-INF/ECLIPSEF.SF" => MergeStrategy.discard
-  case x if x.startsWith("META-INF/services") => MergeStrategy.concat
-  case _ => MergeStrategy.first
-}
-
-PB.targets in Compile := Seq(
-  scalapb.gen() -> (sourceManaged in Compile).value
-)
