--- geopyspark-backend/vectorpipe/src/main/scala/geopyspark/vectorpipe/package.scala
@@ -1,4 +1,0 @@
-package geopyspark
-
-
-package object vectorpipe extends protobufs.Implicits
