--- geopyspark-backend/vectorpipe/src/main/scala/geopyspark/vectorpipe/protobufs/Implicits.scala
@@ -1,6 +1,0 @@
-package geopyspark.vectorpipe.protobufs
-
-
-object Implicits extends Implicits
-
-trait Implicits extends FeatureProtoBuf
