+++ geopyspark/__init__.py
@@ -149,11 +149,8 @@ __all__ = ['geopyspark_conf']
 
 from . import geotrellis
 from .geotrellis import *
-from . import vector_pipe
-from .vector_pipe import *
 from . import geotools
 from .geotools import *
 
 __all__.extend(geotrellis.__all__)
-__all__.extend(vector_pipe.__all__)
 __all__.extend(geotools.__all__)
